Fibonacci last trial

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,15 +6,15 @@
 int main(void)
 {
 	int i;
-	long long int a, b, c;
+	long a, b, c;
 
 	a = 1;
 	b = 2;
-	printf("%lld, %lld, ", a, b);
+	printf("%ld, %ld, ", a, b);
 	for (i = 1; i < 97; i++)
 	{
 		c = a + b;
-		printf("%lld, ", c);
+		printf("%ld, ", c);
 		a = b;
 		b = c;
 	}
